# kb-rag-deploy/docs/openapi/kb-server.yaml
openapi: 3.0.3
info:
  title: kb-rag-server API
  description: 'kb-rag 管理台内部 API（M1 基线 + M2–M8 增量）。


    本文件按 docs/M1-CONTRACTS.md §5、docs/M2-CONTRACTS.md §1.5/§3/§4/§6 与

    docs/M3-CONTRACTS.md §3 定型的端点与 DTO 编写，用于 kb-rag-server / kb-rag-web

    之间的接口契约测试与前端 mock。

    M2 增量：search 新入参（改写/融合/重排/阈值/metadata_filter）与出参（applied、

    各路归一化分）、ik 词典 CRUD、知识库索引配置更新与按新配置重建。ik 词典热更新

    通道 `GET /internal/dict/ik/{ext|stop}.txt`（见 M2-CONTRACTS.md §3）面向 ES ik

    插件而非管理台前端，免鉴权、纯文本响应，不纳入本文件的管理台 API 契约。

    M3 增量：model-status 新增 vision_* 三字段；解析预览与确认（PENDING_CONFIRM 状态、

    单个/批量确认、改规则重解析）；聊天记录导入（匹配预览 + 确认执行）；告警 Webhook

    配置与测试；Demo 一键导入与状态查询；IndexConfig 新增 clean_rules/parse_preview_required/

    chat_aggregation。kb-parser 侧 VLM 调用与图片/清洗均在 kb-rag-server 完成，parser 不

    调用任何大模型（docs/M3-CONTRACTS.md §0/§1）。

    M4a 增量：文档版本列表与激活影响预检（`GET .../versions`、`GET .../activate-impact`、
    `POST .../activate`）、跨版本待复核标注清单、分片标注四操作（编辑/启禁用/合并/拆分）。

    M4b 增量（docs/M4b-CONTRACTS.md）：评测体系全链路——评测集与 case 的增删改查、
    证据复核工作台（`stale-cases`/`recheck`）、Demo 示例评测集导入（`import-demo`）、
    检索调试页一键收进评测集（`cases/from-retrieval`）与检索结果反馈标注
    （`retrieval-feedback`）、评测运行配置矩阵与费用预估（`runs`/`runs/estimate`）、
    run 详情/命中明细下钻/多 run 对比（`eval-runs/{runId}`、`.../results`、
    `eval-runs/compare`）；`ActivateImpact.affected_eval_case_count` 由恒 0 占位改为
    真实统计。对外 API Key 网关（应用发布渠道）属于 M4c 范畴，不在本文件中。
    M4c 增量（docs/M4c-CONTRACTS.md）：应用中心全链路——应用 CRUD（`/api/v1/apps`）、版本创建/
...
    输出与候选白名单求交集，空交集/解析失败/超时/未配置模型一律降级检索全部关联库并记
...

...

...
servers:
...
security:
- bearerAuth: []
paths:
...
  /api/v1/auth/oidc/login:
    get:
      tags:
      - auth
      summary: 发起 OIDC 登录（302 到 IdP 授权端点，M16）
      description: 免认证；签发一次性 state（10 分钟过期）后 302 到发现文档中的授权端点。协议未启用时 400。
      security: []
      responses:
        '302':
          description: 重定向到身份提供方
        '400':
          $ref: '#/components/responses/InvalidParam'
  /api/v1/auth/oidc/callback:
    get:
      tags:
      - auth
      summary: OIDC 回调（M16）
      description: '免认证；验 state → code 换 token → 验 id_token（JWKS 验签 + iss/aud/exp）→ JIT
        建号（落默认租户）→ 302 到 {web-base-url}/login#sso_token={token}；失败同样 302，以
        #sso_error={中文原因} 交付——令牌与错误都走 fragment，不进服务器日志与 Referer。'
      security: []
      parameters:
      - name: code
        in: query
        required: false
        schema: {type: string}
      - name: state
        in: query
        required: false
        schema: {type: string}
      responses:
        '302':
          description: 重定向回控制台登录页，结果在 URL fragment
  /api/v1/auth/saml/login:
    get:
      tags:
      - auth
      summary: 发起 SAML 登录（SP-initiated，302 到 IdP SSO URL，M16）
      description: 免认证；携 deflate+base64 的 AuthnRequest 与一次性 RelayState。协议未启用时 400。
      security: []
      responses:
        '302':
          description: 重定向到身份提供方
        '400':
          $ref: '#/components/responses/InvalidParam'
  /api/v1/auth/saml/acs:
    post:
      tags:
      - auth
      summary: SAML 断言消费端点（IdP POST 回来的 Response，M16）
      description: 免认证；验 RelayState → 验 Response 签名（配置的 IdP 证书）→ 验
        NotOnOrAfter/Audience/InResponseTo → 取 NameID 统一落地；结果同样以 302 + fragment 交付。
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                SAMLResponse:
                  type: string
                  description: base64 编码的 SAML Response（字段名由 SAML POST binding 固定）
                RelayState:
                  type: string
      responses:
        '302':
          description: 重定向回控制台登录页，结果在 URL fragment
  /api/v1/auth/cas/login:
    get:
      tags:
      - auth
      summary: 发起 CAS 登录（302 到 {server-url}/login，M16）
      description: 免认证；无自建 state——CAS ticket 本身一次性且绑定 service URL。协议未启用时 400。
      security: []
      responses:
        '302':
          description: 重定向到 CAS 服务器
        '400':
          $ref: '#/components/responses/InvalidParam'
  /api/v1/auth/cas/callback:
    get:
      tags:
      - auth
      summary: CAS 回调（M16）
      description: 免认证；持 ticket 服务端 GET {server-url}/p3/serviceValidate → 解析 cas:user 统一落地；
        结果以 302 + fragment 交付。
      security: []
      parameters:
      - name: ticket
        in: query
        required: false
        schema: {type: string}
      responses:
        '302':
          description: 重定向回控制台登录页，结果在 URL fragment
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: '自定义 Header：Authorization: Bearer <token>。服务端 token 表/内存缓存维护，

        有效期 24h；除 /api/v1/auth/login、/api/v1/auth/sso-available、/api/v1/auth/sso/**（含

        oidc/saml/cas 的 login 与回调）与 /actuator/** 外全部拦截。

        M15 起认证之后还有一道功能权限拦截：命中 @RequiresPermission 且主体不持有其中任一权限码时返回 403 FORBIDDEN。

        '
    ApiKeyBearer:
      type: http
      scheme: bearer
      bearerFormat: kb-sk-*
      description: '对外开放 API（/api/v1/knowledge/*）专用鉴权，Header：Authorization: Bearer

        kb-sk-*；按 key_hash（SHA-256）查验，与管理台 bearerAuth 会话令牌相互独立、走独立过滤器链，

        独立限流（按 Key 令牌桶）（M4c-CONTRACTS.md §3）。

        '
...
  responses:
    InvalidParam:
      description: 参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorCode:
      type: string
      description: M1 错误码子集 + M4c 对外开放 API 新增错误码 + M15 FORBIDDEN，HTTP 状态码见括注
      enum:
      - INVALID_PARAM
      - UNAUTHORIZED
      - FORBIDDEN
      - NOT_FOUND
      - INVALID_API_KEY
      - API_KEY_DISABLED
      - APP_ACCESS_DENIED
      - APP_NOT_FOUND
      - VERSION_NOT_FOUND
      - VERSION_NOT_PUBLISHED
      - RATE_LIMITED
      - PARSE_FAILED
      - UPSTREAM_MODEL_ERROR
      - INTERNAL_ERROR
    ErrorResponse:
      type: object
      description: 统一错误响应
      required:
      - code
      - message
      - request_id
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          example: 'invalid parameter: name is required'
        request_id:
          type: string
          description: 全链路透传的请求 ID，经 X-Request-Id Header 贯穿 kb-server/kb-parser/模型
            Provider 日志
          example: req-3f6a2e1c
...